Subject: [PATCH] CacheFiles: Add a culling index
Add an index to make culling more efficient.  The entries in the index consist
of file handles as obtained through exportfs from the backing filesystem (*not*
from the netfs) and atimes.

Each index entry points to a backing file in the cache by means of the FH, and
the backing file points to the index entry in its attached xattr.

The index is split between two files.  The first file contains the FH's, the
second file contains an array of atimes, whereby the atime slots in the second
file correspond on a one-to-one basis with the index entry slots.

The FH index file is arranged so that (1) all the FH's are the same size, (2)
FH's do not cross page boundaries.  The backing fs is expected to produce FH's
that are all the same size and type for files within the cache.  This avoids
the need to store FH sizes and types in the entry.

CULLING IN USERSPACE
====================

We can still implement the actual culling algorithm in userspace.  Userspace
can scan the atime table looking for files that haven't been accessed for a
while.  The slot index can be used as a key to communicate with the kernel to
find out whether the file is actually in use.

Userspace should delete the backing file first, and then erase the index slot.

CONSISTENCY PROBLEMS
====================

Now, because there are two separated pieces of metadata for each backing file,
these can get out of sync, particularly if the cache is not cleanly shut down.
This can leave the cache in a number of bad states:

 (1) A backing file can end up with no assigned entry, and thus not be seen by
     the culling algorithm.

 (2) An entry can remain uncleared after its backing file has been removed.

 (3) Two backing files can end up pointing to the same entry.

 (4) Incorrect atimes.

In the case of (1), this might be a good way to implement pinned objects as
they won't be seen be the culling algorithm.  However, in the case of normal
operation, this is generally bad.  There are a couple of ways of dealing with
it:

 (a) Have userspace do a background (priority 19) scan when the cache is
     started, looking for backing files that don't match their entries and
     delete them.

 (b) As (a), but make a note that the cache has been sync'd on shutdown.  If
     the note is there on cache startup, delete the note and skip the scan.

In the case of (2), the entry will eventually float to the top of the culling
pile and be deleted when the culling algorithm has to make more space.
However, it might be worth scanning for dud entries here too, as they use up
space that could be allocated to new objects.

In the case of (3), if a backing file points to an entry that doesn't point
back to it, it should be given a new entry.

In the case of (4), we just don't worry about it; this is a cache, not a proper
data store.

Signed-off-by: David Howells <dhowells@redhat.com>